Fixed B field equations and bugs.
removed the 1/D factor in the amplitude of the B field phasor correction for distance because the amplitude of the near field is inside the B field interaction matrix. 
Removed a bug located in the stimation of the initial impedance loading for each element. It was used the noninitisised current array instead of the skin effect stimation.
removed a bug when computing the proximity effect constant for the bases.